Refactor dirs source
